+++ pyupdate/__init__.py
@@ -1,4 +1,5 @@
 from pyupdate.update import UpdateManager
+import pyupdate.utilities as util
 
 
 def info():
