+++ pyupdate/utilities/__init__.py
@@ -1,4 +1,3 @@
 import pyupdate.utilities.hashing as hashing
 import pyupdate.utilities.helper as helper
 from pyupdate.utilities.build import Builder
-from pyupdate.utilities.web import GitManager
